#217 introduce service validates the runtime config before updating it
