+++ src/main/java/kostka/moviecatalog/service/validation/UniqueRuntimeConfigNameValidator.java
@@ -1,7 +1,7 @@
 package kostka.moviecatalog.service.validation;
 
 import kostka.moviecatalog.annotation.UniqueRuntimeConfigName;
-import kostka.moviecatalog.service.runtimeconfiguration.RuntimeConfigurationService;
+import kostka.moviecatalog.service.runtimeconfiguration.RuntimeValidationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,11 +10,11 @@ import javax.validation.ConstraintValidatorContext;
 
 @Service
 public class UniqueRuntimeConfigNameValidator implements ConstraintValidator<UniqueRuntimeConfigName, String> {
-    private final RuntimeConfigurationService runtimeConfigurationService;
+    private final RuntimeValidationService runtimeValidationService;
 
     @Autowired
-    public UniqueRuntimeConfigNameValidator(final RuntimeConfigurationService runtimeConfigurationService) {
-    this.runtimeConfigurationService = runtimeConfigurationService;
+    public UniqueRuntimeConfigNameValidator(final RuntimeValidationService runtimeValidationService) {
+    this.runtimeValidationService = runtimeValidationService;
     }
 
     /**
@@ -26,8 +26,8 @@ public class UniqueRuntimeConfigNameValidator implements ConstraintValidator<Uni
     @Override
     public boolean isValid(final String configName, final ConstraintValidatorContext context) {
         try {
-            runtimeConfigurationService.getConfigByType(
-                    runtimeConfigurationService.getTypeByName(configName)
+            runtimeValidationService.getConfigByType(
+                    runtimeValidationService.getTypeByName(configName)
             );
             return false;
         } catch (Exception e) {
